+++ packages/react/src/avatar/docs/overview.mdx
@@ -6,12 +6,6 @@ storybookPath: /story/Content-Avatar--basic
 figmaGalleryNodeId: 16226%3A40311
 ---
 
-<ResponsivePreview
-	href="/example-site/category/subcategory/multi-page-form"
-	label="Preview viewing a record in a table"
-	title="Responsive record table preview"
-/>
-
 Avatars are used to show a thumbnail representation of an individual.
 
 ```jsx live
@@ -52,7 +46,7 @@ AgDS foreground components respond to the background colour of their parent cont
 
 This logic ensures sufficient contrast between foreground and background elements is maintained to meet WCAG 2.1 AA 4.5:1 contrast ratio for text (WCAG 1.4.3) and 3:1 for graphic elements that convey information (WCAG 1.4.11).
 
-```jsx live previewHeading="Avatar"
+```jsx live
 <>
 	<Box padding={1.5}>
 		<Avatar name="William Mead" tone="action" />
