+++ packages/react/src/table/docs/overview.mdx
@@ -8,166 +8,6 @@ relatedComponents: ['summary-list']
 relatedPatterns: ['selectable-table-with-batch-actions', 'search-filters']
 ---
 
-<ResponsivePreview
-	padding={false}
-	href="/yourgov"
-	label="Preview using href"
-	title="Yourgov preview"
-/>
-
-```jsx previewHeading="Preview using code"
-<TableWrapper>
-	<Table>
-		<TableCaption>Applications (with actions dropdown)</TableCaption>
-		<TableHead>
-			<TableRow>
-				<TableHeader scope="col">Reference</TableHeader>
-				<TableHeader scope="col">Date submitted</TableHeader>
-				<TableHeader scope="col">Status</TableHeader>
-				<TableHeader scope="col">Actions</TableHeader>
-			</TableRow>
-		</TableHead>
-		<TableBody>
-			<TableRow>
-				<TableCell as="th" scope="row" fontWeight="bold">
-					<TextLink href="#" id="REF-ABC3CD4EF">
-						REF-ABC3CD4EF
-					</TextLink>
-				</TableCell>
-				<TableCell>20/06/2024</TableCell>
-				<TableCell>
-					<StatusBadge
-						appearance="subtle"
-						tone="infoMedium"
-						label="In progress"
-					/>
-				</TableCell>
-				<TableCell>
-					<DropdownMenu>
-						<DropdownMenuButton aria-describedby="REF-ABC3CD4EF">
-							Action
-						</DropdownMenuButton>
-						<DropdownMenuPanel>
-							<DropdownMenuItem>Edit</DropdownMenuItem>
-							<DropdownMenuItem>Download</DropdownMenuItem>
-							<DropdownMenuItem>Delete</DropdownMenuItem>
-						</DropdownMenuPanel>
-					</DropdownMenu>
-				</TableCell>
-			</TableRow>
-			<TableRow>
-				<TableCell as="th" scope="row" fontWeight="bold">
-					<TextLink href="#" id="REF-5GHC6IJ7K">
-						REF-5GHC6IJ7K
-					</TextLink>
-				</TableCell>
-				<TableCell>25/06/2024</TableCell>
-				<TableCell>
-					<StatusBadge
-						appearance="subtle"
-						tone="infoMedium"
-						label="In progress"
-					/>
-				</TableCell>
-				<TableCell>
-					<DropdownMenu>
-						<DropdownMenuButton aria-describedby="REF-5GHC6IJ7K">
-							Action
-						</DropdownMenuButton>
-						<DropdownMenuPanel>
-							<DropdownMenuItem>Edit</DropdownMenuItem>
-							<DropdownMenuItem>Download</DropdownMenuItem>
-							<DropdownMenuItem>Delete</DropdownMenuItem>
-						</DropdownMenuPanel>
-					</DropdownMenu>
-				</TableCell>
-			</TableRow>
-			<TableRow>
-				<TableCell as="th" scope="row" fontWeight="bold">
-					<TextLink href="#" id="REF-M8NCO9PQR">
-						REF-M8NCO9PQR
-					</TextLink>
-				</TableCell>
-				<TableCell>02/07/2024</TableCell>
-				<TableCell>
-					<StatusBadge
-						appearance="subtle"
-						tone="successMedium"
-						label="Completed"
-					/>
-				</TableCell>
-				<TableCell>
-					<DropdownMenu>
-						<DropdownMenuButton aria-describedby="REF-M8NCO9PQR">
-							Action
-						</DropdownMenuButton>
-						<DropdownMenuPanel>
-							<DropdownMenuItem>Edit</DropdownMenuItem>
-							<DropdownMenuItem>Download</DropdownMenuItem>
-							<DropdownMenuItem>Delete</DropdownMenuItem>
-						</DropdownMenuPanel>
-					</DropdownMenu>
-				</TableCell>
-			</TableRow>
-			<TableRow>
-				<TableCell as="th" scope="row" fontWeight="bold">
-					<TextLink href="#" id="REF-S1TCU2VWX">
-						REF-S1TCU2VWX
-					</TextLink>
-				</TableCell>
-				<TableCell>05/08/2024</TableCell>
-				<TableCell>
-					<StatusBadge
-						appearance="subtle"
-						tone="infoMedium"
-						label="In progress"
-					/>
-				</TableCell>
-				<TableCell>
-					<DropdownMenu>
-						<DropdownMenuButton aria-describedby="REF-S1TCU2VWX">
-							Action
-						</DropdownMenuButton>
-						<DropdownMenuPanel>
-							<DropdownMenuItem>Edit</DropdownMenuItem>
-							<DropdownMenuItem>Download</DropdownMenuItem>
-							<DropdownMenuItem>Delete</DropdownMenuItem>
-						</DropdownMenuPanel>
-					</DropdownMenu>
-				</TableCell>
-			</TableRow>
-			<TableRow>
-				<TableCell as="th" scope="row" fontWeight="bold">
-					<TextLink href="#" id="REF-Y3ZCA4B5C">
-						REF-Y3ZCA4B5C
-					</TextLink>
-				</TableCell>
-				<TableCell>19/10/2024</TableCell>
-				<TableCell>
-					<StatusBadge
-						appearance="subtle"
-						tone="successMedium"
-						label="Completed"
-					/>
-				</TableCell>
-				<TableCell>
-					<DropdownMenu>
-						<DropdownMenuButton aria-describedby="REF-Y3ZCA4B5C">
-							Action
-						</DropdownMenuButton>
-						<DropdownMenuPanel>
-							<DropdownMenuItem>Edit</DropdownMenuItem>
-							<DropdownMenuItem>Download</DropdownMenuItem>
-							<DropdownMenuItem>Delete</DropdownMenuItem>
-						</DropdownMenuPanel>
-					</DropdownMenu>
-				</TableCell>
-			</TableRow>
-		</TableBody>
-	</Table>
-</TableWrapper>
-```
-
 ```jsx live
 <TableWrapper>
 	<Table>
